refactor(cli): split ws module into file
